+++ src/Component.ts
